+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - MV-322 - webinar viewer's count has to be reset on WebRTC init
+- MV-323 - Webinar note redirect fix
 
 ## [2.86.1] - 2021-03-31
 ### Fixed